Sacrifice one HTTP request for simplicity
The previous architecture would return the command result in the
very response body. This is fast, but complicates the code because
there are two ways to get a response (/run and /outputs). I've
decided to sacrifice this optimization for now and only return
results if requested at /outputs. This will also simplify the
future globbing feature.